Remove the possibility of adding columns with table overrides (except ALIAS columns)
Supporting this properly requires more comprehensive changes.
