+++ src/Interpreters/applyTableOverride.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <Parsers/ASTCreateQuery.h>
+#include <Parsers/ASTIdentifier.h>
+#include <Parsers/ASTTableOverrides.h>
+
+namespace DB
+{
+
+class ASTTableOverride;
+class ASTCreateQuery;
+class ASTIndentifier;
+
+void applyTableOverrideToCreateQuery(const ASTTableOverride & override, ASTCreateQuery * create_query);
+
+}
